one api page for all search
